fix(cli): preserve unverified tag immutability
