+++ packages/cli/src/daemon/auto-update.ts
@@ -44,6 +44,7 @@ import {
   type DkgConfig,
   type ResolvedAutoUpdateConfig,
 } from '../config.js';
+import { parseTagName, resolveAutoUpdateGitRef } from '../auto-update-ref.js';
 import {
   _autoUpdateIo,
   DAEMON_EXIT_CODE_RESTART,
@@ -79,40 +80,6 @@ export function normalizeRepo(repo: string): string {
   const m = t.match(/github\.com[/:](\S+\/\S+?)(?:\/|$)/);
   if (m) return m[1];
   return t;
-}
-
-export function parseTagName(ref: string): string | null {
-  const m = ref.match(/^refs\/tags\/(.+)$/);
-  return m ? m[1] : null;
-}
-
-export function isValidRef(ref: string): boolean {
-  if (!/^[\w./+\-]+$/.test(ref)) return false;
-  if (!ref || ref.startsWith("-") || ref.startsWith("/") || ref.endsWith("/")) return false;
-  if (ref.includes("..") || ref.includes("//") || ref.includes("@{")) return false;
-  if (ref.endsWith(".")) return false;
-  const parts = ref.split("/");
-  return parts.every((part) => {
-    if (!part || part === "." || part === "..") return false;
-    if (part.endsWith(".lock")) return false;
-    return true;
-  });
-}
-
-export function normalizeGitRefInput(ref: string): string {
-  const trimmed = ref.trim() || "main";
-  if (!isValidRef(trimmed)) {
-    throw new Error(`invalid branch/ref "${ref}"`);
-  }
-  if (trimmed.startsWith("refs/")) return trimmed;
-  return `refs/heads/${trimmed}`;
-}
-
-export function resolveAutoUpdateGitRef(
-  au: Pick<ResolvedAutoUpdateConfig, "branch"> & { ref?: string },
-  refOverride?: string,
-): string {
-  return normalizeGitRefInput(refOverride ?? au.ref ?? au.branch);
 }
 
 export function isValidRepoSpec(repo: string): boolean {
@@ -1243,7 +1210,7 @@ async function _performUpdateInner(
 
   try {
     const maybeTag = parseTagName(ref);
-    const fetchRef = maybeTag ? `+${ref}:${ref}` : ref;
+    const fetchRef = maybeTag && verifyTagSignature ? `+${ref}:${ref}` : ref;
     const fetchStartedAt = Date.now();
     log(
       `Auto-update (git): fetching "${ref}" from ${fetchUrl} into slot ${target}...`,
